dashboard crashing on file move
