+++ src/components/emptyBagView.js
