+++ src/components/itemInfoView.js
@@ -10,25 +10,25 @@ const DATA = [
         id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
         name: 'Blueberries',
         quantity: '250g',
-        image: require('../assets/chef.jpg')
+        image: require('../assets/img/chef.jpg')
     },
     {
         id: '3ac68afc-c605-48d3-a4f8-fbd91aa97f63',
         name: 'Orange',
         quantity: '1pcs',
-        image: require('../assets/chef2.jpg')
+        image: require('../assets/img/chef2.jpg')
     },
     {
         id: '58694a0f-3da1-471f-bd96-145571e29d72',
         name: 'Egg',
         quantity: '1pcs',
-        image: require('../assets/chef1.jpg')
+        image: require('../assets/img/chef1.jpg')
     },
     {
         id: '58684a0f-3da1-471f-bd96-145571e29d72',
         name: 'Chocolate',
         quantity: '150g',
-        image: require('../assets/chef3.jpg')
+        image: require('../assets/img/chef3.jpg')
     },
 ];
