+++ src/components/navBarView.js
