+++ src/home/home.data.js
@@ -1,0 +1,44 @@
+export const FEATURESDATA = [
+    {
+        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
+        image: require('../assets/inventory.png'),
+        title: 'Fast delivery',
+        description: 'And free returns. Contact us for more info.',
+        image: require('../assets/icon/delivery.png')
+    },
+    {
+        id: '3ac68afc-c605-48d3-a4f8-fbd91aa97f63',
+        image: require('../assets/inventory.png'),
+        title: 'Manage your inventory',
+        description: 'Add or remove from your inventory.',
+    },
+    {
+        id: '58694a0f-3da1-471f-bd96-145571e29d72',
+        image: require('../assets/payment.svg'),
+        title: 'Payment processing',
+        description: 'All payments will be taken care for you.',
+    }
+];
+
+export const DATA = [
+    {
+        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
+        title: 'Bradley Harmon',
+        image: require('../assets/img/chef.jpg')
+    },
+    {
+        id: '3ac68afc-c605-48d3-a4f8-fbd91aa97f63',
+        title: 'Shantelle Bob',
+        image: require('../assets/img/chef2.jpg')
+    },
+    {
+        id: '58694a0f-3da1-471f-bd96-145571e29d72',
+        title: 'Jose Montes',
+        image: require('../assets/img/chef1.jpg')
+    },
+    {
+        id: '58684a0f-3da1-471f-bd96-145571e29d72',
+        title: 'Jordi Becker',
+        image: require('../assets/img/chef3.jpg')
+    },
+];
