+++ src/legal/Legal.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import { registerRootComponent } from 'expo';
 import { StatusBar } from 'expo-status-bar';
-import styles from '../components/style';
+import styles from '../../components/style';
 import { Platform, PixelRatio, Dimensions, Text, View } from 'react-native';
 const {
     height: windowHeight,
