+++ src/order/Order.js
@@ -4,52 +4,9 @@ import { Image, Text, View, SafeAreaView, ScrollView } from 'react-native';
 
 import CustomHeader from '../components/customHeader'
 import { FlatList } from 'react-native-gesture-handler';
+import { DATA } from './order.data.js';
 
-const DATA = [
-    {
-        id: 'bd7acba-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1512621776951-a57141f2eefd?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80'
-    },
-    {
-        id: '0-c605-48d3-a4f8-fbd91aa97f63',
-        name: 'Second Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
 
-    },
-    {
-        id: '0-3da1-471f-bd96-145571e29d72',
-        name: 'Third Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-
-    },
-    {
-        id: 'bd7acbea-0-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    }, {
-        id: 'bdacbea-c1b1-0-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    }, {
-        id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    },
-    {
-        id: '0-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    }, {
-        id: '8-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    }, {
-        id: '6-c1b1-46c2-aed5-3ad53abb28ba',
-        name: 'First Item',
-        imageURL: 'https://images.unsplash.com/photo-1482049016688-2d3e1b311543?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=706&q=80'
-    },
-];
 
 const Item = ({ imageURL, name }) => (
     <View style={{ flexDirection: 'row', height: 70 }}>
